requirements.txt
asgiref==3.8.1
astroid==3.3.10
dill==0.4.0
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
exceptiongroup==1.3.0
Faker==37.4.0
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
migrate==0.3.8
model-bakery==1.20.5
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
python-decouple==3.8
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.3
typing_extensions==4.14.0
tzdata==2025.2
